+++ src/main/java/codewars/tasks/java/action/LeapYearsExecution.java
@@ -1,0 +1,12 @@
+package codewars.tasks.java.action;
+
+import codewars.tasks.java.leapYears.LeapYears;
+
+public class LeapYearsExecution {
+
+    public static void main (String [] args){
+
+        LeapYears isYearLeap = new LeapYears();
+        isYearLeap.isYearLeap(2004);
+    }
+}
