+++ src/main/java/codewars/tasks/java/action/TemperatureAnalysisExecution.java
@@ -1,0 +1,11 @@
+package codewars.tasks.java.action;
+
+import codewars.tasks.java.temperatureAnalysis.TemperatureAnalysis;
+
+public class TemperatureAnalysisExecution {
+
+    public static void main (String [] args){
+
+        TemperatureAnalysis analysis = new TemperatureAnalysis();
+    }
+}
