+++ src/main/java/codewars/tasks/java/action/Vector3DExecution.java
@@ -1,0 +1,12 @@
+package codewars.tasks.java.action;
+
+import codewars.tasks.java.vector.Vector3D;
+
+public class Vector3DExecution {
+
+    public static void main (String [] args){
+
+        Vector3D vector = new Vector3D();
+        vector.getLength(0,2,0);
+    }
+}
